create delete modal for re use component

+++ src/components/dealerTable/DealerTable.jsx
@@ -1,38 +1,60 @@
 "use client"
-import Link from 'next/link'
+import { toast } from 'react-toastify';
 import React, { useState } from 'react'
 import { FaRegEye, FaPencilAlt } from "react-icons/fa";
 import { ImBin } from "react-icons/im";
 import SingleDealerView from '../SingleDealerView/SingleDealerView';
+import DeleteAlart from '../deleteConfirmationAlert/DeleteAlart';
 
 function DealerTable({ dealers }) {
 
-    const [selectDealer, setSelectSealer] = useState(null)
+    const [selectDealer, setSelectDealer] = useState(null)
+    const [isModalOpen, setIsModalOpen] = useState(false)
 
     const handleView = (dealerData) => {
-        setSelectSealer(dealerData)
+        setSelectDealer(dealerData)
         document.getElementById("my_modal_5").showModal();
     }
 
-    const handleDleteDelar = (id) => {
-        console.log(id)
-        fetch(`http://127.0.0.1:8000/dealers/${id}/`, {
+    const handleDeleteClick = (dealer) => {
+        console.log("Dealer Button clicked")
+        setSelectDealer(dealer);
+        setIsModalOpen(true);
+    }
+
+
+    const handleConfirmDelete = () => {
+
+        if (selectDealer) {
+            console.log("Deleting dealer:", selectDealer.id);
+            fetch(`http://127.0.0.1:8000/dealers/${selectDealer.id}/`, {
             method: 'DELETE',
         })
 
-        // .then(res => res)
-        // .then(() =>{
-        //      toast.success("Dealer deleted successfully")
-        // })
+            .then(res => res)
+            .then(() => {
+                toast.success("Dealer deleted successfully")
+            })
 
-        // .catch(error => {
-        //     toast.error(error.message);
-        //     setLoading(false)
-        // })
+            .catch(error => {
+                toast.error(error.message);
+                setLoading(false)
+            })
+        }
+        setIsModalOpen(false);
+
+
+        
+
+
     }
 
     return (
         <div>
+
+
+
+
             <div className='my-6'>
                 <h3 className='md:text-2xl text-xl mb-2 text-[#57c7d4]'>🧾Dealer List</h3>
                 <div className="overflow-x-auto rounded-box border border-base-content/5 bg-base-100 shadow-lg">
@@ -62,9 +84,10 @@ function DealerTable({ dealers }) {
                                             <button onClick={() => { handleView(deale) }} className="badge badge-soft badge-info mr-2"><span className='text-xl'><FaRegEye /></span>
                                             </button>
 
-                                            <button onClick={() => { handleDleteDelar(deale.id) }} className="badge badge-soft text-red-400 mr-2">
+                                            <button onClick={() => { handleDeleteClick(deale) }} className="badge badge-soft text-red-400 mr-2">
                                                 <span><ImBin></ImBin></span>
                                             </button>
+                                            <DeleteAlart id={deale.id} status='dealer'></DeleteAlart>
 
                                             <button className="badge badge-soft badge-primary mr-2">
                                                 <span> <FaPencilAlt></FaPencilAlt> </span></button>
@@ -77,6 +100,11 @@ function DealerTable({ dealers }) {
                 </div>
             </div>
             <SingleDealerView dealer={selectDealer}></SingleDealerView>
+            <DeleteAlart
+                isOpen={isModalOpen}
+                onClose={() => setIsModalOpen(false)}
+                onConfirm={handleConfirmDelete}
+            />
         </div>
     )
 }
